helper functions are added under the utils file
